+++ zebra-network/src/peer_set/initialize/tests/vectors.rs
@@ -34,7 +34,7 @@ use crate::{
     address_book_updater::AddressBookUpdater,
     config::CacheDir,
     constants, init,
-    meta_addr::{MetaAddr, PeerAddrState},
+    meta_addr::MetaAddr,
     peer::{self, ClientTestHarness, HandshakeRequest, OutboundConnectorRequest},
     peer_set::{
         initialize::{
@@ -1227,33 +1227,18 @@ async fn self_connections_should_fail() {
     // Wait until the crawler has tried at least one self-connection
     tokio::time::sleep(TEST_CRAWL_NEW_PEER_INTERVAL * 3).await;
 
-    // Check that the self-connection failed
+    // Check that the self-connection failed and the proven self address was suppressed.
     let self_connection_status = {
         let mut unlocked_address_book = address_book
             .lock()
             .expect("unexpected panic in address book");
 
-        let self_connection_status = unlocked_address_book
-            .get(real_self_listener.addr())
-            .expect("unexpected dropped listener address in address book");
-
-        std::mem::drop(unlocked_address_book);
-
-        self_connection_status
+        unlocked_address_book.get(real_self_listener.addr())
     };
 
-    // Make sure we fetched from the address book correctly
     assert_eq!(
-        self_connection_status.addr(),
-        real_self_listener.addr(),
-        "wrong address fetched from address book"
-    );
-
-    // Make sure the self-connection failed
-    assert_eq!(
-        self_connection_status.last_connection_state,
-        PeerAddrState::Failed,
-        "self-connection should have failed"
+        self_connection_status, None,
+        "proven self address should be removed from the address book"
     );
 }